rebase and fix sub
