+++ src/zkevm_specs/opcode/__init__.py
@@ -1,4 +1,4 @@
-from .add import *
+from .add_sub import *
 from .byte import *
 from .comparator import *
 from .lt_gt import *
